Fix up cuda compilation

+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/grid/GridPointers.java
@@ -77,7 +77,7 @@ public class GridPointers {
     // Op extraArgs
     private Pointer extraArgs;
 
-    public GridPointers(Op op, int... dimensions) {
+    public GridPointers(Op op, long... dimensions) {
         this.type = BaseOp.getOpType(op);
         this.dtype = op.x().data().dataType();
         this.opNum = op.opNum();
