+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/impl/meta/ReduceMetaOp.java
@@ -34,7 +34,7 @@ public class ReduceMetaOp extends BaseMetaOp {
         super();
     }
 
-    public ReduceMetaOp(ScalarOp opA, ReduceOp opB, int... dimensions) {
+    public ReduceMetaOp(ScalarOp opA, ReduceOp opB, long... dimensions) {
         this(new OpDescriptor(opA), new OpDescriptor(opB, dimensions));
     }
